+++ resources/views/documents/add.blade.php
@@ -13,11 +13,23 @@
 
             <div class="form-group form-floating mb-3">
                 <input type="text" class="form-control" name="name" value="{{ old('username') }}" placeholder="Title" required="required" autofocus>
+              
                 <label for="floatingName">Title</label>
                 @if ($errors->has('name'))
                     <span class="text-danger text-left">{{ $errors->first('name') }}</span>
                 @endif
             </div>
+            <div class="form-group form-floating mb-3">
+                <input type="text" class="form-control" name="short_description" value="{{ old('short_description') }}" placeholder="Enter a short description" required="required" autofocus>
+
+              
+                <label for="floatingName">Short description</label>
+                @if ($errors->has('short_description'))
+                    <span class="text-danger text-left">{{ $errors->first('short_description') }}</span>
+                @endif
+            </div>
+
+
             
             <textarea name="content" id="myeditorinstance"></textarea>