+++ resources/views/documents/view.blade.php
@@ -3,7 +3,21 @@
 @section('content')
     <div class="bg-light p-5 rounded">
         @auth
-            {{$doc->name}}
+            <div class="document__view">
+                <h1 class="text-primary">
+                    {{$doc->name}}
+                </h1>
+                <small class="text-secondary">
+                    {{$doc->short_description}}
+
+                </small>
+            </div>
+            <br><br>
+            <h5>Content:</h5>
+            <div id="document__content container">
+                {!! html_entity_decode($doc->content) !!}
+            </div>
+
 
         @endauth