+++ resources/views/home/results.blade.php
@@ -4,16 +4,25 @@
     <div class="bg-light p-5 rounded">
         @auth
         @include('home.partials.search')
+        <small>Total results: </small>
 
-        <ul>
+        <hr>
+     
             @foreach ($response as $r )
-                <li>{{$r['_source']['name']}}</li>
+                {{-- <li>{{$r['_source']['name']}}</li> --}}
+                <a class="results__link" href="/doc/view/{{$r['_id']}}">
+                    <div class="card w-100 mt-4 border-0" style="width: 18rem;">
+                        <div class="card-body results__card__body ">
+                            <div>
+                              <h5 class="card-title text-primary">{{$r['_source']['name']}}</h5>
+                              <small class="text-secondary">{{$r['_source']['short_description']}}</small>
+    
+                            </div>
+                        </div>
+                      </div>
+                </a>
+
             @endforeach
-        </ul>
-
-
-
-      
       
         @endauth
 
@@ -23,3 +32,18 @@
         @endguest
     </div>
 @endsection
+
+<style>
+    .results__card__body{
+        display: flex;
+        
+    }
+    .results__card__body  .btn__view{
+        position: absolute;
+        right: 2%;
+    }
+    .results__card__body a, .results__link{
+        text-decoration: none;
+    }
+
+</style>