+++ CHANGELOG.md
@@ -9,3 +9,11 @@
     - Added font awesome for icons library
     - Added register and login system
     - Added layouts for general app and auth.
+
+
+# 0.1.0
+- Added support for document search
+- Added support for document view page
+- Fix Documents controller after adding document command call.
+- Added short_description to documentation table
+- Implement short_description add documentation